+++ cmd/ganymede/main.go
