+++ services/callisto/mysql/riskcontrol/account_risk_ strategy.go
@@ -1,4 +1,4 @@
-package amusingriskcontrol
+package riskcontrol
 
 import (
 	"amusingx.fit/amusingx/mysqlstruct/amusingriskcontrol"
