+++ services/callisto/mysql/riskcontrol/account_risk_ strategy_test.go
@@ -1,4 +1,4 @@
-package amusingriskcontrol
+package riskcontrol
 
 import (
 	"amusingx.fit/amusingx/services/callisto/conf"
