+++ services/callisto/mysql/riskcontrol/verification_code_risk_test.go
@@ -1,4 +1,4 @@
-package amusingriskcontrol
+package riskcontrol
 
 import (
 	"amusingx.fit/amusingx/mysqlstruct/amusingriskcontrol"
